<p dir="auto">TL;DR; Hot path (e.g. counter increment) was calling atomicModifyIORef two times. This can become a bottleneck in certain scenarios.</p>
<ol dir="auto">
<li>Use double-locking for a lock-free optimistic case – no need to lock on the map if the metric already exits. A map lookup was done inside atomic op previously.</li>
<li>Use atomic counter instead of IORef for counters for lock-free operation. Under the hood, they use atomic increment operation that is implemented by modern CPUs.</li>
</ol>

+++ instrument/test/src/Instrument/Tests/Client.hs
@@ -117,7 +117,7 @@ time_test mkConn step = do
         let getCount payload =
               case payload of
                 (AggStats stats) -> scount stats
-                (AggCount n) -> n
+                (AggCount n) -> fromIntegral n
 
             countMatches name n = do
               void $ step name
